Write JSDoc comments for main modules.
This is an attempt to see how easy / hard it would be to write out all
our types as JSDoc comments and then generate our “*.d.ts” files from
those doc-strings.

This way, the main modules _remain actually runnable in a browser_, and
TypeScript tooling can still be leveraged for better development.